requirements.txt
discord.py
discord-pretty-help
beautifulsoup4
schedule